Rename some internal types for brevity
